+++ components/elements/Command.tsx
@@ -4,7 +4,7 @@ import { DialogProps } from "@radix-ui/react-dialog";
 import { Command as CommandPrimitive } from "cmdk";
 
 import { cn } from "../util";
-import { Dialog, DialogContent } from "./Dialog";
+import { Dialog, DialogContent } from "./dialog/Dialog";
 
 const Command = React.forwardRef<
   React.ElementRef<typeof CommandPrimitive>,